+++ R/rast.R
@@ -418,10 +418,10 @@ setMethod("rast", signature(x="ANY"),
 		xyz <- as.matrix(xyz)
 		xyz <- matrix(as.numeric(xyz), ncol=ncol(xyz), nrow=nrow(xyz))
 	}
-	xyz[,1] <- round(xyz[,1], digits)
-	xyz[,2] <- round(xyz[,2], digits)
 	
 	x <- sort(unique(xyz[,1]))
+#	x <- sort(unique(round(xyz[,1], digits+2)))
+	
 	if (length(x) == 1) {
 		error("rast", "cannot create a raster geometry from a single x coordinate")
 	}
@@ -440,13 +440,11 @@ setMethod("rast", signature(x="ANY"),
 	}
 
 	y <- sort(unique(xyz[,2]))
+#	y <- sort(unique(round(xyz[,2], digits+2)))
 	if (length(y) == 1) {
 		error("rast", "cannot create a raster geometry from a single y coordinate")
 	}
 	dy <- y[-1] - y[-length(y)]
-	# probably a mistake to use the line below
-	# Gareth Davies suggested that it be removed
-	# dy <- round(dy, digits)
 
 	ry <- min(dy)
 	for (i in 1:5) {