+++ NEWS.md
@@ -13,7 +13,7 @@
 
 - `as.array<SpatRasterDataset>` method
 - `distance<SpatRaster,missing>` now has argument "values". If TRUE, the values of the nearest non-target cell is returned instead of the distance [#1243](https://github.com/rspatial/terra/issues/1243) by Simon Dedman
-- `threshold<SpatRaster>` [#1233](https://github.com/rspatial/terra/issues/1233) by Agustin Lobo
+- `twoClasses<SpatRaster>` [#1233](https://github.com/rspatial/terra/issues/1233) by Agustin Lobo
 
 # version 1.8-15